requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
PyMuPDF==1.23.8
camelot-py==0.11.0
pytesseract==0.3.10
Pillow==10.1.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
reportlab==4.0.9
jinja2==3.1.2
